+++ frontend/src/components/Layout/Header.jsx
@@ -1,4 +1,4 @@
-import { Link } from 'react-router-dom'
+import { Link } from 'react-router-dom';
 
 function Header() {
   return (
@@ -7,14 +7,20 @@ function Header() {
         <h1>Tikettijärjestelmä</h1>
         <nav>
           <ul>
-            <li><Link to="/">Tiketit</Link></li>
-            <li><Link to="/new-ticket">Luo tiketti</Link></li>
-            <li><Link to="/login">Kirjaudu</Link></li>
+            <li>
+              <Link to="/">Tiketit</Link>
+            </li>
+            <li>
+              <Link to="/new-ticket">Luo tiketti</Link>
+            </li>
+            <li>
+              <Link to="/login">Kirjaudu</Link>
+            </li>
           </ul>
         </nav>
       </div>
     </header>
-  )
+  );
 }
 
-export default Header 
+export default Header;